Components creation and comments
