+++ tvscheduler-react-app/tvscheduler-react-app/src/pages/mainSchedulePage.css
@@ -3,83 +3,3 @@
     flex-direction: column;
     gap: 10px;
 }
-
-span.line{
-  background: linear-gradient(180deg, #adadad85 20%, #a9a9a9b5 65%, #b1b1b18a 75%, #adadad85 0) 100% / 100% 500% no-repeat;
-  position: absolute;
-  top: 40%;
-  left: 8%;
-  display: block;
-  height: 15px;
-  width: 142px;
-  border-radius: 3px;
-  animation: loading 2s ease-in-out infinite;
-  
-  &:nth-child(1){
-    top: 72%;
-    width: 195px;
-    height: 8px;
-  }
-  
-  &:nth-child(2){
-    top: 65%;
-    width: 262px;
-    height: 8px;
-  }
-}
-
-@keyframes loading{
-    0% {
-    background-position: 0 0; 
-  }
-  100% {
-    background-position: 0 100%;
-  }
-}
-
-.dummy-show{
-  box-sizing: border-box;
-  overflow:hidden;
-  position: relative;
-  background: #1c1c1c;
-  border: 2px solid #121212;
-  width: 360px;
-  height: 200px;
-  border-radius: 5px;
-  color: white;
-  z-index: 1;
-
-  & h3{
-    white-space: nowrap;
-    margin-block: 0;
-    position: absolute;
-    top: 40%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-  }
-
-  & p{
-    position: absolute;
-    top: 68%;
-    left: 14%;
-  }
-
-  & span{
-    position: absolute;
-    margin-left: 2.5%;
-
-    & img{
-      transform: translateY(20%);
-    }
-  }
-
-  & .add-button{
-    cursor: default;
-    &:hover{
-      color: white
-    }
-    &::before, &::after{
-      display: none;
-    }
-  }
-}
